Tools.js renamed to ToolList.js and made to use CollapsePanel logic
