+++ client/src/stylesheets/base/_global.scss
@@ -62,6 +62,10 @@ $information-phone: 'only screen and (max-width : 800px)';
   background-color: $green-color;
 }
 
+.bg-blue-color {
+  background-color: $main-blue;
+}
+
 .float-right {
   float: right;
 }
