+++ client/src/stylesheets/main.scss
@@ -30,7 +30,7 @@
   'components/_info',
   'components/Card',
   'components/Comment',
-  'components/Tools',
+  'components/ToolList',
   'components/NotFound';
 
 //Do not change the order of these imports:
